requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
clr-loader==0.2.6
comtypes==1.2.1
contourpy==1.2.0
cycler==0.12.1
darkdetect==0.8.0
dj-database-url==2.1.0
Django==5.0.1
djangorestframework==3.14.0
docker==7.0.0
fonttools==4.47.2
gunicorn==21.2.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pillow==10.2.0
psutil==5.9.8
psycopg==3.1.17
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
whitenoise==6.6.0
